feat(api/events routes): add Endpoint GET events/?filter=...
