+++ api/src/controllers/events.ts
@@ -1,13 +1,17 @@
 import { Request, Response, NextFunction } from "express"
-import { getEventsFromDb, getEventsFromDbBySearch } from '../utils/utilsEvents'
+import { getEventsFromDb, getEventsFromDbBySearch, getEventsFromDbByFilter } from '../utils/events'
 
 export const getEvents = async (req: Request, res: Response, next: NextFunction) => {
     try {
         const search = req.query.search as string
+        const filter = req.query.filter as string
         let events;
         if (search) {
             events = await getEventsFromDbBySearch(search.trim())
-        } else {
+        } 
+        if(filter){
+            events = await getEventsFromDbByFilter(filter)
+        }else {
             events = await getEventsFromDb()
         }
         res.status(200).json(events)
